+++ packages/rockets-server-auth/src/domains/role/application/queries/handlers/abstract-rockets-get-roles-by-ids.handler.ts
@@ -5,7 +5,7 @@ import {
   getDynamicRepositoryToken,
   type RepositoryInterface,
   Where,
-} from '@concepta/nestjs-repository';
+} from '@concepta/rockets-core';
 
 import { ROLE_CRUD_ENTITY_KEY } from '../../../../../shared/constants/repository-entity-keys.constants';
 import { RocketsGetRolesByIdsQuery } from '../impl/rockets-get-roles-by-ids.query';