+++ packages/rockets-server-auth/src/shared/authentication/rockets-auth-password-port.handlers.ts
@@ -1,4 +1,4 @@
-import { AppContextHost } from '@concepta/rockets-core';
+import { AppContextHost, TransactionScope } from '@concepta/rockets-core';
 import {
   CommandBus,
   CommandHandler,
@@ -17,7 +17,6 @@ import {
 } from '@concepta/nestjs-user';
 import { ValidateCurrentPasswordCommand } from '@concepta/nestjs-password';
 import { EventContextHost } from '@concepta/nestjs-core';
-import { TransactionScope } from '@concepta/nestjs-repository';
 
 import { GetActiveCredentialQuery } from '../../domains/user/application/queries/impl/get-active-credential.query';
 import {